+++ src/library/policies/policy.c
@@ -60,7 +60,7 @@
 #define POLICY_INFO   2
 #define POLICY_PHASES 2
 #define POLICY_GPU    2
-#define POLICY_SAVINGS_LVL 0
+#define POLICY_SAVINGS_LVL 2
 
 #define DEF_USE_CPUPRIO 0
 
@@ -522,24 +522,13 @@ void fill_cpus_out_of_cgroup(uint exc)
     int cpu_cnt, cpu_idx;
     if (exc == 0) return;
     cpu_cnt = arch_desc.top.cpu_count;
-    #if DLB
-    verbosen_master(1, "IDLE freq (%lu) cpus --> ",
-                    frequency_pstate_to_freq(frequency_get_num_pstates()-1));
-    #endif
 
     for (cpu_idx = 0; cpu_idx < cpu_cnt; cpu_idx++){
         if (freq_per_core[cpu_idx] == 0){
             debug("Reducing frequency of CPU %d", cpu_idx);
             freq_per_core[cpu_idx] = frequency_pstate_to_freq(frequency_get_num_pstates()-1);
-            #if DLB
-            verbosen_master(1,"%d ", cpu_idx);
-            #endif
         }
     }
-    #if DLB
-    verbosen_master(1,"\n");
-    #endif
-
 }
 
 #define DEBUG_CPUFREQ_COST 0
@@ -1200,7 +1189,7 @@ state_t policy_node_apply(signature_t *my_sig, ulong *freq_set, int *ready)
       if (earl_phase_classification == APP_CPU_GPU)
       {
             // verbose_master(POLICY_PHASES,"%sCPU-GPU application%s", COL_BLU,COL_CLR);
-            verbose_master(1, "%sCPU-GPU application%s (%s)", COL_BLU, COL_CLR, node_name);
+            verbose_master(2, "%sCPU-GPU application%s (%s)", COL_BLU, COL_CLR, node_name);
 
             // TODO: The below function does not modify avg_nf, but it's used just after the function call.
             st = policy_cpu_gpu_freq_selection(c, my_sig, freqs, &freqs_domain, &avg_nf);
@@ -1847,7 +1836,7 @@ state_t policy_mpi_end(mpi_call call_type)
 
         st = polsyms_fun.mpi_end(c,call_type, &per_process_node_freq, &process_id);
         if (state_fail(st)) {
-            verbose(0, "%sERROR%s at policy MPI call end: %s", COL_RED, COL_CLR, state_msg);
+            verbose(2, "%sERROR%s at policy MPI call end: %s", COL_RED, COL_CLR, state_msg);
         }
 
         if (process_id != NO_PROCESS) {
@@ -2032,7 +2021,7 @@ static state_t from_proc_to_core(ulong *process_cpu_freqs_khz, int process_cnt, 
         int sem_ret_val = sem_wait(lib_shared_lock_sem);
         if (sem_ret_val < 0)
         {
-            verbose(1, "%sWARNING%s Locking semaphor for cpu mask read failed: %d", COL_YLW, COL_CLR, errno);
+            verbose(2, "%sWARNING%s Locking semaphor for cpu mask read failed: %d", COL_YLW, COL_CLR, errno);
         }
 
         cpu_set_t process_mask = sig_shared_region[proc_lrank].cpu_mask;
@@ -2042,7 +2031,7 @@ static state_t from_proc_to_core(ulong *process_cpu_freqs_khz, int process_cnt, 
         {
             if (sem_post(lib_shared_lock_sem) < 0)
             {
-                verbose(1, "%sWARNING%s Unlocking semaphor for cpu mask read failed: %d",
+                verbose(2, "%sWARNING%s Unlocking semaphor for cpu mask read failed: %d",
                         COL_YLW, COL_CLR, errno);
             }
         }